Refactor : 컴포넌트 의존성 분리
- todo를 구성하는 각 컴포넌트가 하나의 관심사만 처리하도록 컴포넌트 분할
- depth가 깊어지는 문제가 있어 Context를 통한 상태 관리
